add error treatment byID

+++ src/main/java/me/tiagoluz/forum/controller/TopicoController.java
@@ -16,6 +16,7 @@ import org.springframework.web.util.UriComponentsBuilder;
 import javax.validation.Valid;
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/topicos")
@@ -50,23 +51,38 @@ public class TopicoController {
   }
 
   @GetMapping("/{id}")
-  public DetalhesTopicoDto detail(@PathVariable Long id) {
-    Topico topico = topicoRepository.getOne(id);
-    return new DetalhesTopicoDto(topico);
+  public ResponseEntity<DetalhesTopicoDto> detail(@PathVariable Long id) {
+    Optional<Topico> topico = topicoRepository.findById(id);
+    if (topico.isPresent()) {
+      return ResponseEntity.ok(new DetalhesTopicoDto(topico.get()));
+    }
+    return ResponseEntity.notFound().build();
+
+//    return topico.map(value ->
+//      ResponseEntity.ok(new DetalhesTopicoDto(value))).orElseGet(() ->
+//      ResponseEntity.notFound().build());
   }
 
   @PutMapping("/{id}")
   @Transactional
   public ResponseEntity<TopicoDto> update(@PathVariable Long id, @RequestBody @Valid UpdateTopicForm form) {
-    Topico topico = form.update(id, topicoRepository);
-    return ResponseEntity.ok(new TopicoDto(topico));
+    Optional<Topico> optional = topicoRepository.findById(id);
+    if (optional.isPresent()) {
+      Topico topico = form.update(id, topicoRepository);
+      return ResponseEntity.ok(new TopicoDto(topico));
+    }
+    return ResponseEntity.notFound().build();
   }
 
   @DeleteMapping("/{id}")
   @Transactional
   public ResponseEntity<?> delete(@PathVariable Long id) {
-    topicoRepository.deleteById(id);
-    return ResponseEntity.ok().build();
+    Optional<Topico> optional = topicoRepository.findById(id);
+    if (optional.isPresent()) {
+      topicoRepository.deleteById(id);
+      return ResponseEntity.ok().build();
+    }
+    return ResponseEntity.notFound().build();
   }
 
 }
